added all the required changes
