+++ .gitignore
@@ -1,2 +1,5 @@
 build/*
 .cache/*
+DS_Store/*
+.vscode/settings.json
+
