+++ CMakeLists.txt
@@ -37,7 +37,7 @@ endif()
 include_directories(src)
 
 # NOTE: add source files for your players here
-set(player_src src/player/human_player.cpp src/player/my_player.cpp src/player/my_observer.cpp)
+set(player_src tests/human_player.cpp src/player/my_player.cpp src/player/my_observer.cpp)
 add_library(tttplayer STATIC ${player_src})
 target_link_libraries(tttplayer ${TTTCORE_LIB})
