+++ README.md
@@ -106,6 +106,18 @@
   игроками; `cli_client` — исполняемый файл, который можно собрать с кодом
   вашего игрока для игры на удаленном сервере.
 
+### Игрок-человек для тестирования
+
+В папке tests реализован класс HumanPlayer, который позволяет:
+
+   - Играть против базовых игроков (легкого и сложного), чтобы понять их стратегии
+   - Тестировать своего бота в интерактивном режиме для отладки и наблюдения за его поведением
+   - Наглядно видеть состояние игрового поля после каждого хода
+   
+Использование: запустите test_my_player_vs_human после сборки проекта. По умолчанию ваш игрок настроен как X, а человек — как O. Вы можете изменить конфигурацию, раскомментировав соответствующие строки в файле test_my_player_vs_human.cpp.
+
+Это дополнение к автоматическим тестам существенно упрощает разработку и отладку вашего алгоритма.
+
 ### Сборка проекта
 
 Сборка базовой версии проекта:
