--- src/core/utils.hpp
@@ -1,11 +1,0 @@
-#pragma once
-
-#include "game.hpp"
-#include <iostream>
-
-namespace ttt::utils {
-
-
-void print_game_state(const ttt::game::State& state);
-
-} // namespace ttt::utils
