+++ tests/human_player.cpp
