+++ tests/test_baseline.cpp
@@ -29,48 +29,13 @@ int main(int argc, char *argv[]) {
   ttt::game::Game game(opts);
   game.add_player(ttt::game::Sign::X, p1);
   game.add_player(ttt::game::Sign::O, p2);
-  // game.add_player(ttt::game::Sign::O, &prand);
   game.add_observer(&obs);
 
   
   while (game.process() == ttt::game::MoveResult::OK) {
-    //can't call the function here either because needs to be in libtttcore.a ?
-    //ttt::game::print_game_state(game.get_state());
+  
+    obs.print_game_state(game.get_state());
     
-    // Print the current state of the board
-    std::cout << "   "; //extra space for column index
-    for (int x = 0; x < opts.cols; ++x) {
-      std::cout << std::setw(2) << x%10;
-    }
-    std::cout << "\n";
-    
-    //line separator
-    std::cout << "   +";
-    for (int x = 0; x < opts.cols; ++x) {
-      std::cout << "--";
-    }
-    std::cout << "\n";
-    
-    // print board with row indexes
-    for (int y = 0; y < opts.rows; ++y) {
-      std::cout << std::setw(2) << y << " |";
-      for (int x = 0; x < opts.cols; ++x) {
-        char c = '.';
-        switch (game.get_state().get_value(x, y)) {
-        case ttt::game::Sign::X:
-          c = 'X';
-          break;
-        case ttt::game::Sign::O:
-          c = 'O';
-          break;
-        default:
-          break;
-        }
-        std::cout << c << " "; //add spaces for readability and better debugging
-      }
-      std::cout << "\n";
-    }
-    std::cout << "\n";
     
   }
